Test Cases for SD
When we run the test cases, it will not run them as JUNIT for some
reason. We are still debugging our config issues. Once these issues are
resolved, the tests should be able to work. It will not even fail a test
assertEquals(1, 0). Getting classNotFound exception on the class itself.